fix(tar-xz): address Copilot round 4 findings on PR #108
R4-1 (M/correctness): mode defaults used '||' which treats mode 0
(legitimate TAR entry) as falsy, silently replacing with default. Use
'??' so only undefined/null triggers the default. 3 call sites in
node/file.ts; node/create.ts and browser/create.ts already used '??'.

R4-2 (M/correctness): extractFile applied 'strip' to entry.name but
not to entry.linkname for HARDLINK entries. With strip:1, archive
'dir/link → dir/a.txt' would fail because 'dir/a.txt' linkname still
includes the stripped 'dir/' prefix. Apply strip to linkname before
resolve() and re-run path safety check. Regression test added.

R4-3 (L/doc): README claimed preset 3 was 'Browser default' but actual
default is 6 (uniform across Node and Browser). Doc updated.

R4-4 (L/dry): collectAllChunks/decompressXz/runWritable were duplicated
between list.ts and extract.ts. Extracted to internal node/xz-helpers.ts
to prevent drift.
