+++ packages/tar-xz/src/node/list.ts
@@ -2,77 +2,12 @@
  * Node.js TAR listing with XZ decompression — v6 AsyncIterable API
  */
 
-import { Readable, Writable } from 'node:stream';
-import { createUnxz } from 'node-liblzma';
+import { Writable } from 'node:stream';
 import { calculatePadding } from '../tar/index.js';
-import { toAsyncIterable, type TarInputNode } from '../internal/to-async-iterable.js';
+import type { TarInputNode } from '../internal/to-async-iterable.js';
 import type { TarEntry } from '../types.js';
 import { type HeaderParserState, parseNextHeader } from './tar-parser.js';
-
-// ---------------------------------------------------------------------------
-// Internal helpers (mirrors node/extract.ts — kept local to avoid circular dep)
-// ---------------------------------------------------------------------------
-
-async function collectAllChunks(input: TarInputNode): Promise<Uint8Array> {
-  const iterable = toAsyncIterable(input);
-  const chunks: Uint8Array[] = [];
-  for await (const chunk of iterable) {
-    chunks.push(chunk);
-  }
-  const total = chunks.reduce((n, c) => n + c.length, 0);
-  const out = new Uint8Array(total);
-  let offset = 0;
-  for (const chunk of chunks) {
-    out.set(chunk, offset);
-    offset += chunk.length;
-  }
-  return out;
-}
-
-async function decompressXz(data: Uint8Array): Promise<Uint8Array> {
-  const unxzStream = createUnxz();
-  const readable = Readable.from(
-    (async function* () {
-      yield data;
-    })()
-  );
-
-  const output: Uint8Array[] = [];
-  let resolveFlush!: () => void;
-  let rejectFlush!: (e: unknown) => void;
-  const done = new Promise<void>((res, rej) => {
-    resolveFlush = res;
-    rejectFlush = rej;
-  });
-
-  unxzStream.on('data', (...args: unknown[]) => {
-    const chunk = args[0] as Buffer;
-    output.push(new Uint8Array(chunk.buffer, chunk.byteOffset, chunk.byteLength));
-  });
-  unxzStream.on('end', resolveFlush);
-  unxzStream.on('error', rejectFlush);
-  readable.pipe(unxzStream);
-
-  await done;
-
-  const total = output.reduce((n, c) => n + c.length, 0);
-  const result = new Uint8Array(total);
-  let offset = 0;
-  for (const chunk of output) {
-    result.set(chunk, offset);
-    offset += chunk.length;
-  }
-  return result;
-}
-
-async function runWritable(writable: Writable, data: Uint8Array): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    writable.on('finish', resolve);
-    writable.on('error', reject);
-    writable.write(Buffer.from(data.buffer, data.byteOffset, data.byteLength));
-    writable.end();
-  });
-}
+import { collectAllChunks, decompressXz, runWritable } from './xz-helpers.js';
 
 /**
  * Writable stream that lists TAR entries without extracting content
